fix: Change GalleryItem ordering field to i64
Updated the type of the ordering field in the GalleryItem struct from Option<i32> to Option<i64> to support larger integer values.

+++ src/models/modrinth/entities.rs
@@ -28,7 +28,7 @@ pub struct GalleryItem {
     pub description: Option<String>,
     #[serde(deserialize_with = "chrono_datetime_as_bson_datetime::deserialize")]
     pub created: DateTime<Utc>,
-    pub ordering: Option<i32>,
+    pub ordering: Option<i64>,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone, ToSchema)]
